Removes multiprocessing as it didn't work in Python3
